Extract FindPartitionSplitPoint
